Removed --usedefault of mysql_secure_installation from init_server. Added NO_USE_TLS environment var.

+++ deploy_files/server_deploy.sh
@@ -24,7 +24,9 @@ sudo rm /etc/nginx/sites-enabled/default
 sudo cp nginx/sites-enabled/default /etc/nginx/sites-enabled/
 
 # Configure tls
-chmod +x configure_tls.sh
-./configure_tls.sh
-ps aux | grep nginx | grep -v grep | awk '{ print "kill ", $2 }' | sh # certbot nginx plugin bug
-sudo systemctl stop nginx
+if [ $NO_USE_TLS != "True" ]; then
+  chmod +x configure_tls.sh
+  ./configure_tls.sh
+  ps aux | grep nginx | grep -v grep | awk '{ print "kill ", $2 }' | sh # certbot nginx plugin bug
+  sudo systemctl stop nginx
+fi
